requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv

google-adk[a2a]
google-genai

google-cloud-aiplatform>=1.60.0
google-cloud-storage>=2.16.0

mcp

httpx